re-implemented NUTS
