+++ desilike/samplers/__init__.py
@@ -1,18 +1,17 @@
 """Module providing wrappers for commonly used samplers."""
 
+from .blackjax import HMCSampler, NUTSSampler
 from .dynesty import DynestySampler
 from .emcee import EmceeSampler
 from .grid import GridSampler
-from .hmc import HMCSampler
 from .importance import ImportanceSampler
 # from .mclmc import MCLMCSampler
 # from .mcmc import MCMCSampler
 from .nautilus import NautilusSampler
-# from .nuts import NUTSSampler
 from .pocomc import PocoMCSampler
 from .qmc import QMCSampler
 from .zeus import ZeusSampler
 
 __all__ = ['DynestySampler', 'GridSampler', 'HMCSampler', 'EmceeSampler',
-           'ImportanceSampler', 'NautilusSampler', 'PocoMCSampler',
-           'QMCSampler', 'ZeusSampler']
+           'ImportanceSampler', 'NautilusSampler', 'NUTSSampler',
+           'PocoMCSampler', 'QMCSampler', 'ZeusSampler']
